refactor(project): import instance context directly
