+++ packages/opencode/src/effect/instance-ref.ts
@@ -1,5 +1,5 @@
 import { Context } from "effect"
-import type { InstanceContext } from "@/project/instance"
+import type { InstanceContext } from "@/project/instance-context"
 import type { WorkspaceID } from "@/control-plane/schema"
 
 export const InstanceRef = Context.Reference<InstanceContext | undefined>("~opencode/InstanceRef", {
