+++ packages/opencode/src/format/formatter.ts
@@ -1,5 +1,5 @@
 import { Npm } from "@opencode-ai/core/npm"
-import type { InstanceContext } from "../project/instance"
+import type { InstanceContext } from "../project/instance-context"
 import { Filesystem } from "@/util/filesystem"
 import { Process } from "@/util/process"
 import { which } from "../util/which"
